requirements.txt
PySide6==6.9.0
pyqtgraph
pyserial
numpy
scipy